Move autodiscovery script target name script to common
+ Update webshell-common-ts submodule
